@@ -1,10 +1,24 @@
 package jp.linanfine.dsma.dialog;
 
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.Intent;
+import android.os.Handler;
+import android.util.Log;
+import android.view.View;
+import android.webkit.WebChromeClient;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
+
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.TreeMap;
 
-import jp.linanfine.dsma.activity.ScoreList;
+import jp.linanfine.dsma.R;
 import jp.linanfine.dsma.util.common.TextUtil;
 import jp.linanfine.dsma.util.file.FileReader;
 import jp.linanfine.dsma.value.GateSetting;
@@ -15,277 +29,239 @@ import jp.linanfine.dsma.value.UniquePattern;
 import jp.linanfine.dsma.value._enum.FullComboType;
 import jp.linanfine.dsma.value._enum.MusicRank;
 import jp.linanfine.dsma.value._enum.PatternType;
-import jp.linanfine.dsma.R;
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.app.AlertDialog;
-import android.content.DialogInterface;
-import android.content.Intent;
-import android.os.Handler;
-import android.util.Log;
-import android.view.View;
-import android.webkit.WebChromeClient;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-import android.widget.LinearLayout;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
 
+/**
+ * 曲リスト表示後の[GATEサーバから一括取得(詳細)]関連のコード、のはず
+ */
 public class DialogFromGateSou {
-	
-	public static int LoginRequestCode = 20003;
-	
-	private Handler mHandler = new Handler();
-	private Activity mParent;
-	private AlertDialog mDialog;
-	private View mView;
-	
-	//private TreeMap<Integer, MusicData> mMusicList;
-	private IdToWebMusicIdList mWebMusicIds;
-	private TreeMap<Integer, MusicScore> mScoreList;
-	private WebView mWebView;
-	private ProgressBar mWebProgress;
-	private GateSetting mGateSetting;
-	private ArrayList<UniquePattern> mSouList;
-	private String mRivalId;
-	private String mRivalName;
-	
-	private int mPageCount = 0;
-	private int mCurrentPage = 0;
-	private TextView mLogView = null;
-	private TextView mLogView2 = null;
-	private ProgressBar mProgress = null;
-	private TextView mPercent = null;
-	private TextView mCurrent = null;
-	private TextView mMax = null;
-	
-	private String mUriH;
-	private String mUriM;
-	private String mUriF;
-	private String mRequestUri;
-	private int mItemId;
-	private String mWebItemId;
-	private PatternType mPattern;
 
-	private boolean mCanceled = false;
-	
-	@SuppressLint("SetJavaScriptEnabled")
-	public DialogFromGateSou(Activity parent) {
-		mParent = parent;
-		mView = mParent.getLayoutInflater().inflate(R.layout.view_from_gate_list, null);
+    public static int LoginRequestCode = 20003;
+
+    private final Handler mHandler = new Handler();
+    private final Activity mParent;
+    private AlertDialog mDialog;
+    private final View mView;
+
+    //private TreeMap<Integer, MusicData> mMusicList;
+    private IdToWebMusicIdList mWebMusicIds;
+    private TreeMap<Integer, MusicScore> mScoreList;
+    private WebView mWebView;
+    private ProgressBar mWebProgress;
+    private GateSetting mGateSetting;
+    private ArrayList<UniquePattern> mSouList;
+    private String mRivalId;
+    private String mRivalName;
+
+    private int mPageCount = 0;
+    private int mCurrentPage = 0;
+    private TextView mLogView = null;
+    private TextView mLogView2 = null;
+    private ProgressBar mProgress = null;
+    private TextView mPercent = null;
+    private TextView mCurrent = null;
+    private TextView mMax = null;
+
+    private String mUriH;
+    private String mUriM;
+    private String mUriF;
+    private String mRequestUri;
+    private int mItemId;
+    private String mWebItemId;
+    private PatternType mPattern;
+
+    private boolean mCanceled = false;
+
+    @SuppressLint("SetJavaScriptEnabled")
+    public DialogFromGateSou(Activity parent) {
+        mParent = parent;
+        mView = mParent.getLayoutInflater().inflate(R.layout.view_from_gate_list, null);
 
         Intent intent = mParent.getIntent();
-        if(intent == null)
-        {
-        	return;
+        if (intent == null) {
+            return;
         }
-        
+
         //mMusicList = FileReader.readMusicList(mParent);
-		mGateSetting = FileReader.readGateSetting(mParent);
-        
+        mGateSetting = FileReader.readGateSetting(mParent);
+
         mWebMusicIds = FileReader.readWebMusicIds(mParent);
 
-        mWebProgress = (ProgressBar)mView.findViewById(R.id.webProgress);
-		mLogView = (TextView)mView.findViewById(R.id.log);
-		mLogView2 = (TextView)mView.findViewById(R.id.log2);
-		mPercent = (TextView)mView.findViewById(R.id.percent);
-		mCurrent = (TextView)mView.findViewById(R.id.current);
-		mMax = (TextView)mView.findViewById(R.id.max);
-		mProgress = (ProgressBar)mView.findViewById(R.id.mainProgress);
-		
-		mLogView2.setVisibility(View.VISIBLE);
-		mView.findViewById(R.id.log3).setVisibility(View.VISIBLE);
-        
-        WebViewClient client = new WebViewClient() 
-        {
+        mWebProgress = mView.findViewById(R.id.webProgress);
+        mLogView = mView.findViewById(R.id.log);
+        mLogView2 = mView.findViewById(R.id.log2);
+        mPercent = mView.findViewById(R.id.percent);
+        mCurrent = mView.findViewById(R.id.current);
+        mMax = mView.findViewById(R.id.max);
+        mProgress = mView.findViewById(R.id.mainProgress);
+
+        mLogView2.setVisibility(View.VISIBLE);
+        mView.findViewById(R.id.log3).setVisibility(View.VISIBLE);
+
+        WebViewClient client = new WebViewClient() {
             @Override
             public void onPageFinished(WebView view, String url) {
-    			view.loadUrl("javascript:window.viewsourceactivity.viewSource(document.documentElement.outerHTML);");
+                view.loadUrl("javascript:window.viewsourceactivity.viewSource(document.documentElement.outerHTML);");
             }
+
             @Override
             public void onPageStarted(WebView view, String url, android.graphics.Bitmap favicon) {
-            	mWebProgress.setProgress(1);
+                mWebProgress.setProgress(1);
             }
-    	};
-    	
-    	WebChromeClient chrome = new WebChromeClient() 
-    	{
-    		public void onProgressChanged(WebView view, int progress) {
-    			//try { Thread.sleep(10);} catch (InterruptedException e) {}
-    			mWebProgress.setProgress(1+progress);
-    			mProgress.setProgress(mCurrentPage*100+(mPageCount==0?0:progress));
-    			mPercent.setText(String.valueOf((100*mCurrentPage+(mPageCount==0?0:progress))/(mPageCount==0?10000:mPageCount))+"%");
-    		}
-    	};
+        };
 
-    	mWebView = (WebView) mView.findViewById(R.id.webView);
+        WebChromeClient chrome = new WebChromeClient() {
+            public void onProgressChanged(WebView view, int progress) {
+                //try { Thread.sleep(10);} catch (InterruptedException e) {}
+                mWebProgress.setProgress(1 + progress);
+                mProgress.setProgress(mCurrentPage * 100 + (mPageCount == 0 ? 0 : progress));
+                mPercent.setText((100 * mCurrentPage + (mPageCount == 0 ? 0 : progress)) / (mPageCount == 0 ? 10000 : mPageCount) + "%");
+            }
+        };
+
+        mWebView = mView.findViewById(R.id.webView);
         //mWebView.getSettings().setBlockNetworkImage(true);
-    	mWebView.getSettings().setBuiltInZoomControls(true);
-    	mWebView.setWebViewClient(client);
-    	mWebView.setWebChromeClient(chrome);
+        mWebView.getSettings().setBuiltInZoomControls(true);
+        mWebView.setWebViewClient(client);
+        mWebView.setWebChromeClient(chrome);
         //mWebView.getSettings().setJavaScriptEnabled(true);
-    	mWebView.getSettings().setJavaScriptEnabled(true);
-    	mWebView.addJavascriptInterface(this, "viewsourceactivity");
+        mWebView.getSettings().setJavaScriptEnabled(true);
+        mWebView.addJavascriptInterface(this, "viewsourceactivity");
 
-	}
-	
-	public boolean setArguments(AlertDialog dialog, ArrayList<UniquePattern> souList, String rivalId, String rivalName)
-	{
-		mDialog = dialog;
+    }
+
+    public boolean setArguments(AlertDialog dialog, ArrayList<UniquePattern> souList, String rivalId, String rivalName) {
+        mDialog = dialog;
         mRivalId = rivalId;
         mRivalName = rivalName;
         mSouList = souList;
         mScoreList = FileReader.readScoreList(mParent, mRivalId);
         return true;
     }
-	
-	public View getView()
-	{
-		return mView;
-	}
-	
-	private static int getPatternInt(PatternType pattern)
-	{
+
+    public View getView() {
+        return mView;
+    }
+
+    private static int getPatternInt(PatternType pattern) {
 
         return
-    		pattern == PatternType.bSP ? 0 :
-			pattern == PatternType.BSP ? 1 :
-			pattern == PatternType.DSP ? 2 :
-			pattern == PatternType.ESP ? 3 :
-			pattern == PatternType.CSP ? 4 :
-			pattern == PatternType.BDP ? 5 :
-			pattern == PatternType.DDP ? 6 :
-			pattern == PatternType.EDP ? 7 :
-			pattern == PatternType.CDP ? 8 :
-    		0;
-        
-	}
-	
-	public void start()
-	{
-		if(mDialog == null)
-		{
-			return;
-		}
-		FileReader.requestAd((LinearLayout)mView.findViewById(R.id.adContainer), mParent);
-		mPageCount = mSouList.size();
-		if(mPageCount <= 0)
-		{
-			return;
-		}
-    	mProgress.setMax(mPageCount*100);
-    	mMax.setText(String.valueOf(mPageCount));
+                pattern == PatternType.bSP ? 0 :
+                        pattern == PatternType.BSP ? 1 :
+                                pattern == PatternType.DSP ? 2 :
+                                        pattern == PatternType.ESP ? 3 :
+                                                pattern == PatternType.CSP ? 4 :
+                                                        pattern == PatternType.BDP ? 5 :
+                                                                pattern == PatternType.DDP ? 6 :
+                                                                        pattern == PatternType.EDP ? 7 :
+                                                                                pattern == PatternType.CDP ? 8 :
+                                                                                        0;
 
-		mGateSetting = FileReader.readGateSetting(mParent);
-		if(mGateSetting.FromA3)
-		{
-			mUriH = "https://p.eagate.573.jp/game/ddr/ddra3/p/";
-		}
-		else{
-			mUriH = "https://p.eagate.573.jp/game/ddr/ddra20/p/";
-		}
+    }
 
-        if(mRivalId == null)
-        {
-        	mUriH += "playdata/music_detail.html?index=";
-        	mUriM = "&diff=";
-        	mUriF = "";
+    public void start() {
+        if (mDialog == null) {
+            return;
         }
-        else
-        {
-        	mUriH += "rival/music_detail.html?index=";
-        	mUriM = "&diff=";
-        	mUriF = "&rival_id="+mRivalId+"&name="+mRivalName.trim();
+        FileReader.requestAd(mView.findViewById(R.id.adContainer), mParent);
+        mPageCount = mSouList.size();
+        if (mPageCount <= 0) {
+            return;
+        }
+        mProgress.setMax(mPageCount * 100);
+        mMax.setText(String.valueOf(mPageCount));
+
+        mGateSetting = FileReader.readGateSetting(mParent);
+        if (mGateSetting.FromA3) {
+            mUriH = "https://p.eagate.573.jp/game/ddr/ddra3/p/";
+        } else {
+            mUriH = "https://p.eagate.573.jp/game/ddr/ddra20/p/";
+        }
+
+        if (mRivalId == null) {
+            mUriH += "playdata/music_detail.html?index=";
+            mUriM = "&diff=";
+            mUriF = "";
+        } else {
+            mUriH += "rival/music_detail.html?index=";
+            mUriM = "&diff=";
+            mUriF = "&rival_id=" + mRivalId + "&name=" + mRivalName.trim();
         }
         //String uri = "file:///android_asset/status.html";
         mCurrentPage = 0;
         UniquePattern c = mSouList.get(mCurrentPage);
         mItemId = c.MusicId;
-        if(!mWebMusicIds.containsKey(mItemId) || mWebMusicIds.get(mItemId).idOnWebPage == "")
-        {
-    		new AlertDialog.Builder(mParent)
-	        .setIcon(android.R.drawable.ic_dialog_alert)
-	        .setMessage(mParent.getResources().getString(R.string.illegal_id_alert))
-	        .setCancelable(true)
-	        .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), new DialogInterface.OnClickListener() {
-	            public void onClick(DialogInterface dialog, int whichButton) {
-	            	
-	            }
-	        }).show();
-    		mDialog.cancel();
-        	return;
-        }
-        else
-        {
-        	mWebItemId = mWebMusicIds.get(mItemId).idOnWebPage;
+        if (!mWebMusicIds.containsKey(mItemId) || mWebMusicIds.get(mItemId).idOnWebPage.equals("")) {
+            new AlertDialog.Builder(mParent)
+                    .setIcon(android.R.drawable.ic_dialog_alert)
+                    .setMessage(mParent.getResources().getString(R.string.illegal_id_alert))
+                    .setCancelable(true)
+                    .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), (dialog, whichButton) -> {
+                    }).show();
+
+            mDialog.cancel();
+            return;
+        } else {
+            mWebItemId = mWebMusicIds.get(mItemId).idOnWebPage;
         }
         mPattern = c.Pattern;
-        mRequestUri = mUriH+String.valueOf(mWebItemId)+mUriM+String.valueOf(getPatternInt(mPattern))+mUriF;
-        mLogView.setText("Loading...\n"+(mRivalName==null?"My Score\n":("Rival: "+mRivalName+"\n")) + mPattern.toString() + " : " + c.musics.get(mItemId).Name);
+        mRequestUri = mUriH + mWebItemId + mUriM + getPatternInt(mPattern) + mUriF;
+        mLogView.setText("Loading...\n" + (mRivalName == null ? "My Score\n" : ("Rival: " + mRivalName + "\n")) + mPattern.toString() + " : " + c.musics.get(mItemId).Name);
         mWebView.loadUrl(mRequestUri);
-	}
-	
-	public void cancel()
-	{
-		mCanceled = true;
-        WebView web = (WebView) mView.findViewById(R.id.webView);
+    }
+
+    public void cancel() {
+        mCanceled = true;
+        WebView web = mView.findViewById(R.id.webView);
         web.stopLoading();
         mPageCount = 1;
-        if(mDialog != null)
-        {
-        	mDialog.cancel();
+        if (mDialog != null) {
+            mDialog.cancel();
         }
-	}
+    }
 
-	private boolean analyzeScore(String src)
-	{
+    // TODO analyzeScoreをユニットテストでやれるようにUIまわりのコードを分割したい
+    private boolean _analyzeScore(String src) {
+        WebView web = mView.findViewById(R.id.webView);
+        String uri = web.getUrl();
 
-        WebView web = (WebView) mView.findViewById(R.id.webView);
-    	String uri = web.getUrl();
-    	//boolean loggedin = false;
-    	if(!uri.equals(mRequestUri))
-    	{
-        	if(mRivalId == null)
-            {
-        		return false;
+        return analyzeScore(src, uri);
+    }
+
+    private boolean analyzeScore(String src, String uri) {
+
+        //boolean loggedin = false;
+        if (!uri.equals(mRequestUri)) {
+            if (mRivalId == null) {
+                // TODO return false条件を抜き出してメソッド外に置く
+                // 一応、途中まで進めないと行けないパターンがないか探す
+                return false;
+            } else {
+                if (!uri.contains(mRequestUri.split("&name=")[0])) {
+                    return false;
+                }
             }
-            else
-            {
-            	if (!uri.contains(mRequestUri.split("&name=")[0]))
-            	{
-        			return false;
-            	}
+        }
+        ScoreData scoreData = new ScoreData();
+        String cmp = "0\"></td>  <td>";
+        Log.d("", "1");
+        if (src.contains(cmp)) {
+            String dr = src.substring(src.indexOf(cmp) + cmp.length());
+            cmp = "<br>";
+            dr = TextUtil.escapeWebTitle(dr.substring(0, dr.indexOf(cmp)).trim());
+            if (!dr.equals(mWebMusicIds.get(mItemId).titleOnWebPage)) {
+                new AlertDialog.Builder(mParent)
+                        .setIcon(android.R.drawable.ic_dialog_info)
+                        .setMessage(mParent.getResources().getString(R.string.name_different_alert) + "\n\n" + mWebMusicIds.get(mItemId).titleOnWebPage + "\n↓\n" + dr)
+                        .setCancelable(true)
+                        .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), (dialog, whichButton) -> {
+                        })
+                        .show();
+                return true;
             }
-    	}
-    	src = src;
-    	ScoreData sd = new ScoreData();
-		String cmp = "0\"></td>  <td>";
-		Log.d("", "1");
-        if(src.contains(cmp))
-        {
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "<br>";
-        	dr = TextUtil.excapeWebTitle(dr.substring(0, dr.indexOf(cmp)).trim());
-        	if(!dr.equals(mWebMusicIds.get(mItemId).titleOnWebPage))
-        	{
-        		new AlertDialog.Builder(mParent)
-    	        .setIcon(android.R.drawable.ic_dialog_info)
-    	        .setMessage(mParent.getResources().getString(R.string.name_different_alert)+"\n\n"+mWebMusicIds.get(mItemId).titleOnWebPage+"\n↓\n"+dr)
-    	        .setCancelable(true)
-    	        .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), new DialogInterface.OnClickListener() {
-    	            public void onClick(DialogInterface dialog, int whichButton) {
-    	            	
-    	            }
-    	        }).show();
-        		return true;
-        	}
+        } else {
+            return false;
         }
-        else
-        {
-        	return false;
-        }
-		Log.d("", "2");
+        Log.d("", "2");
         cmp = "NO PLAY...";
         //if(src.contains(cmp) || loggedin)
         //{
@@ -303,296 +279,239 @@ public class DialogFromGateSou {
             )).start();
             return true;*/
         //}
-	//else
-        if(!src.contains(cmp))
-    {
-    		Log.d("", "3");
-        if(mCanceled)
-        {
-        	return true;
-        }
-		Log.d("", "4");
-		if(mGateSetting.FromA3)
-		{
-			cmp = "<th>ハイスコア時のランク</th><td>";
-		}
-		else
-		{
-			if (mRivalId == null)
-			{
-				cmp = "<th>ハイスコア時のダンスレベル</th><td>";
-			} else
-			{
-				cmp = "<th>最高ダンスレベル</th><td>";
-			}
-		}
-        if(src.contains(cmp))
-        {
-    		Log.d("", "5");
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	if(dr.equals("AAA"))
-        	{
-        		sd.Rank = MusicRank.AAA;
-        	}
-        	else if(dr.equals("AA+"))
-        	{
-        		sd.Rank = MusicRank.AAp;
-        	}
-        	else if(dr.equals("AA"))
-        	{
-        		sd.Rank = MusicRank.AA;
-        	}
-        	else if(dr.equals("AA-"))
-        	{
-        		sd.Rank = MusicRank.AAm;
-        	}
-        	else if(dr.equals("A+"))
-        	{
-        		sd.Rank = MusicRank.Ap;
-        	}
-        	else if(dr.equals("A"))
-        	{
-        		sd.Rank = MusicRank.A;
-        	}
-        	else if(dr.equals("A-"))
-        	{
-        		sd.Rank = MusicRank.Am;
-        	}
-        	else if(dr.equals("B+"))
-        	{
-        		sd.Rank = MusicRank.Bp;
-        	}
-        	else if(dr.equals("B"))
-        	{
-        		sd.Rank = MusicRank.B;
-        	}
-        	else if(dr.equals("B-"))
-        	{
-        		sd.Rank = MusicRank.Bm;
-        	}
-        	else if(dr.equals("C+"))
-        	{
-        		sd.Rank = MusicRank.Cp;
-        	}
-        	else if(dr.equals("C"))
-        	{
-        		sd.Rank = MusicRank.C;
-        	}
-        	else if(dr.equals("C-"))
-        	{
-        		sd.Rank = MusicRank.Cm;
-        	}
-        	else if(dr.equals("D+"))
-        	{
-        		sd.Rank = MusicRank.Dp;
-        	}
-        	else if(dr.equals("D"))
-        	{
-        		sd.Rank = MusicRank.D;
-        	}
-        	else if(dr.equals("E"))
-        	{
-        		sd.Rank = MusicRank.E;
-        	}
-        	else
-        	{
-        		sd.Rank = MusicRank.Noplay;
-        	}
-        }
-        else
-        {
-        	return false;
-        }
-		Log.d("", "6");
-    	cmp = "<th>ハイスコア</th><td>";
-        if(src.contains(cmp))
-        {
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	sd.Score = Integer.valueOf(dr);
-        }
-        else
-        {
-        	Toast.makeText(mParent, "Failed", Toast.LENGTH_LONG).show();
-        }
-    	cmp = "<th>最大コンボ数</th><td>";
-        if(src.contains(cmp))
-        {
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	sd.MaxCombo = Integer.valueOf(dr);
-        }
-        else
-        {
-        	return false;
-        }
-		Log.d("", "7");
-    	if(mRivalId == null)
-        {
-    	cmp = "<tr><th>フルコンボ種別</th><td>";
-        if(src.contains(cmp))
-        {
-    		Log.d("", "8");
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	if(dr.equals("グッドフルコンボ"))
-        	{
-        		sd.FullComboType = FullComboType.GoodFullCombo;
-        	}
-        	else if(dr.equals("グレートフルコンボ"))
-        	{
-        		sd.FullComboType = FullComboType.FullCombo;
-        	}
-        	else if(dr.equals("パーフェクトフルコンボ"))
-        	{
-        		sd.FullComboType = FullComboType.PerfectFullCombo;
-        	}
-        	else if(dr.equals("マーベラスフルコンボ"))
-        	{
-        		sd.FullComboType = FullComboType.MerverousFullCombo;
-        	}
-        	else
-        	{
-        		sd.FullComboType = FullComboType.None;
-        	}
-        }
-        else
-        {
-        	return false;
-        }
-		Log.d("", "9");
-    	cmp = "<th>プレー回数</th><td>";
-        if(src.contains(cmp))
-        {
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	sd.PlayCount = Integer.valueOf(dr);
-        }
-        else
-        {
-        	return false;
-        }
-		Log.d("", "10");
-    	cmp = "<th>クリア回数</th><td>";
-        if(src.contains(cmp))
-        {
-        	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-        	cmp = "</td>";
-        	dr = dr.substring(0, dr.indexOf(cmp));
-        	sd.ClearCount = Integer.valueOf(dr);
-        }
-        else
-        {
-        	return false;
-        }
-        }
-    	else
-    	{
-        	cmp = "<th>フルコンボ種別</th><td>";
-            if(src.contains(cmp))
-            {
-        		Log.d("", "8");
-            	String dr = src.substring(src.indexOf(cmp)+cmp.length());
-            	cmp = "</td>";
-            	dr = dr.substring(0, dr.indexOf(cmp));
-            	if(dr.equals("グッドフルコンボ"))
-            	{
-            		sd.FullComboType = FullComboType.GoodFullCombo;
-            	}
-            	else if(dr.equals("グレートフルコンボ"))
-            	{
-            		sd.FullComboType = FullComboType.FullCombo;
-            	}
-            	else if(dr.equals("パーフェクトフルコンボ"))
-            	{
-            		sd.FullComboType = FullComboType.PerfectFullCombo;
-            	}
-            	else if(dr.equals("マーベラスフルコンボ"))
-            	{
-            		sd.FullComboType = FullComboType.MerverousFullCombo;
-            	}
-            	else
-            	{
-            		sd.FullComboType = FullComboType.None;
-            	}
+        //else
+        if (!src.contains(cmp)) {
+            Log.d("", "3");
+            if (mCanceled) {
+                return true;
             }
-            else
-            {
-            	return false;
+            Log.d("", "4");
+            if (mGateSetting.FromA3) {
+                cmp = "<th>ハイスコア時のランク</th><td>";
+            } else {
+                if (mRivalId == null) {
+                    cmp = "<th>ハイスコア時のダンスレベル</th><td>";
+                } else {
+                    cmp = "<th>最高ダンスレベル</th><td>";
+                }
             }
-    	}
-	}
-		Log.d("", "11");
-        MusicScore ms;
-        if(mScoreList.containsKey(mItemId))
-        {
-        	ms = mScoreList.get(mItemId);
-        	mScoreList.remove(mItemId);
+            if (src.contains(cmp)) {
+                Log.d("", "5");
+                String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                cmp = "</td>";
+                dr = dr.substring(0, dr.indexOf(cmp));
+                // TODO TextUtilに作ったメソッドに置き換える
+                switch (dr) {
+                    case "AAA":
+                        scoreData.Rank = MusicRank.AAA;
+                        break;
+                    case "AA+":
+                        scoreData.Rank = MusicRank.AAp;
+                        break;
+                    case "AA":
+                        scoreData.Rank = MusicRank.AA;
+                        break;
+                    case "AA-":
+                        scoreData.Rank = MusicRank.AAm;
+                        break;
+                    case "A+":
+                        scoreData.Rank = MusicRank.Ap;
+                        break;
+                    case "A":
+                        scoreData.Rank = MusicRank.A;
+                        break;
+                    case "A-":
+                        scoreData.Rank = MusicRank.Am;
+                        break;
+                    case "B+":
+                        scoreData.Rank = MusicRank.Bp;
+                        break;
+                    case "B":
+                        scoreData.Rank = MusicRank.B;
+                        break;
+                    case "B-":
+                        scoreData.Rank = MusicRank.Bm;
+                        break;
+                    case "C+":
+                        scoreData.Rank = MusicRank.Cp;
+                        break;
+                    case "C":
+                        scoreData.Rank = MusicRank.C;
+                        break;
+                    case "C-":
+                        scoreData.Rank = MusicRank.Cm;
+                        break;
+                    case "D+":
+                        scoreData.Rank = MusicRank.Dp;
+                        break;
+                    case "D":
+                        scoreData.Rank = MusicRank.D;
+                        break;
+                    case "E":
+                        scoreData.Rank = MusicRank.E;
+                        break;
+                    default:
+                        scoreData.Rank = MusicRank.Noplay;
+                        break;
+                }
+            } else {
+                return false;
+            }
+            Log.d("", "6");
+            cmp = "<th>ハイスコア</th><td>";
+            if (src.contains(cmp)) {
+                String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                cmp = "</td>";
+                dr = dr.substring(0, dr.indexOf(cmp));
+                scoreData.Score = Integer.parseInt(dr);
+            } else {
+                Toast.makeText(mParent, "Failed", Toast.LENGTH_LONG).show();
+            }
+            cmp = "<th>最大コンボ数</th><td>";
+            if (src.contains(cmp)) {
+                String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                cmp = "</td>";
+                dr = dr.substring(0, dr.indexOf(cmp));
+                scoreData.MaxCombo = Integer.parseInt(dr);
+            } else {
+                return false;
+            }
+            Log.d("", "7");
+            if (mRivalId == null) {
+                cmp = "<tr><th>フルコンボ種別</th><td>";
+                if (src.contains(cmp)) {
+                    Log.d("", "8");
+                    String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                    cmp = "</td>";
+                    dr = dr.substring(0, dr.indexOf(cmp));
+                    // TODO 同じ処理がいろんなとこにあるはずなのでまとめる
+                    switch (dr) {
+                        case "グッドフルコンボ":
+                            scoreData.FullComboType = FullComboType.GoodFullCombo;
+                            break;
+                        case "グレートフルコンボ":
+                            scoreData.FullComboType = FullComboType.FullCombo;
+                            break;
+                        case "パーフェクトフルコンボ":
+                            scoreData.FullComboType = FullComboType.PerfectFullCombo;
+                            break;
+                        case "マーベラスフルコンボ":
+                            scoreData.FullComboType = FullComboType.MerverousFullCombo;
+                            break;
+                        default:
+                            scoreData.FullComboType = FullComboType.None;
+                            break;
+                    }
+                } else {
+                    return false;
+                }
+                Log.d("", "9");
+                cmp = "<th>プレー回数</th><td>";
+                if (src.contains(cmp)) {
+                    String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                    cmp = "</td>";
+                    dr = dr.substring(0, dr.indexOf(cmp));
+                    scoreData.PlayCount = Integer.parseInt(dr);
+                } else {
+                    return false;
+                }
+                Log.d("", "10");
+                cmp = "<th>クリア回数</th><td>";
+                if (src.contains(cmp)) {
+                    String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                    cmp = "</td>";
+                    dr = dr.substring(0, dr.indexOf(cmp));
+                    scoreData.ClearCount = Integer.parseInt(dr);
+                } else {
+                    return false;
+                }
+            } else {
+                cmp = "<th>フルコンボ種別</th><td>";
+                if (src.contains(cmp)) {
+                    Log.d("", "8");
+                    String dr = src.substring(src.indexOf(cmp) + cmp.length());
+                    cmp = "</td>";
+                    dr = dr.substring(0, dr.indexOf(cmp));
+                    switch (dr) {
+                        case "グッドフルコンボ":
+                            scoreData.FullComboType = FullComboType.GoodFullCombo;
+                            break;
+                        case "グレートフルコンボ":
+                            scoreData.FullComboType = FullComboType.FullCombo;
+                            break;
+                        case "パーフェクトフルコンボ":
+                            scoreData.FullComboType = FullComboType.PerfectFullCombo;
+                            break;
+                        case "マーベラスフルコンボ":
+                            scoreData.FullComboType = FullComboType.MerverousFullCombo;
+                            break;
+                        default:
+                            scoreData.FullComboType = FullComboType.None;
+                            break;
+                    }
+                } else {
+                    return false;
+                }
+            }
         }
-        else
-        {
-        	ms = new MusicScore();
+        Log.d("", "11");
+        MusicScore musicScore;
+        if (mScoreList.containsKey(mItemId)) {
+            musicScore = mScoreList.get(mItemId);
+            mScoreList.remove(mItemId);
+        } else {
+            musicScore = new MusicScore();
         }
-        ScoreData msd;
-        switch(mPattern)
-        {
+        ScoreData localScoreData;
+        switch (mPattern) {
             case bSP:
-            	msd = ms.bSP;
-            	break;
+                localScoreData = musicScore.bSP;
+                break;
             case BSP:
-            	msd = ms.BSP;
-            	break;
+                localScoreData = musicScore.BSP;
+                break;
             case DSP:
-            	msd = ms.DSP;
-            	break;
+                localScoreData = musicScore.DSP;
+                break;
             case ESP:
-            	msd = ms.ESP;
-            	break;
+                localScoreData = musicScore.ESP;
+                break;
             case CSP:
-            	msd = ms.CSP;
-            	break;
+                localScoreData = musicScore.CSP;
+                break;
             case BDP:
-            	msd = ms.BDP;
-            	break;
+                localScoreData = musicScore.BDP;
+                break;
             case DDP:
-            	msd = ms.DDP;
-            	break;
+                localScoreData = musicScore.DDP;
+                break;
             case EDP:
-            	msd = ms.EDP;
-            	break;
+                localScoreData = musicScore.EDP;
+                break;
             case CDP:
-            	msd = ms.CDP;
-            	break;
-        	default:
-        		msd = new ScoreData();
-        		break;
+                localScoreData = musicScore.CDP;
+                break;
+            default:
+                localScoreData = new ScoreData();
+                break;
         }
-        
-        // msd : 元の値
-        // sd  : 取得した値
+
+        // localScoreData(元msd) : 元の値
+        // scoreData(元sd)  : 取得した値
         // 取得した値に元の値を上書きすることによって元の値を維持する
-        
+
         // 「Life4 に未フルコンを上書きする」 が無効
-        if(!mGateSetting.OverWriteLife4)
-        {
-        	// 取得した値が未フルコン
-        	if(sd.FullComboType == FullComboType.None)
-        	{
-            	// 元の値が Life4
-            	if(msd.FullComboType == FullComboType.Life4)
-            	{
-            		// 元のフルコンタイプに戻す
-            		sd.FullComboType = msd.FullComboType;
-            	}
-        	}
+        if (!mGateSetting.OverWriteLife4) {
+            // 取得した値が未フルコン
+            if (scoreData.FullComboType == FullComboType.None) {
+                // 元の値が Life4
+                if (localScoreData.FullComboType == FullComboType.Life4) {
+                    // 元のフルコンタイプに戻す
+                    scoreData.FullComboType = localScoreData.FullComboType;
+                }
+            }
         }
-        
+
         // 「取得したFCでGFCを上書き」 が無効
         /*if(!mGateSetting.OverWriteFullCombo)
         {
@@ -607,23 +526,20 @@ public class DialogFromGateSou {
 	        	}
         	}
         }*/
-        
+
         // 「低いスコアを上書き」 が無効
-        if(!mGateSetting.OverWriteLowerScores)
-        {
-        	// スコアが低かったら
-        	if(sd.Score < msd.Score)
-        	{
-        		// スコアを元に戻す
-        		sd.Score = msd.Score;
-        		sd.Rank = msd.Rank;
-        	}
-        	// コンボが低かったら
-        	if(sd.MaxCombo < msd.MaxCombo)
-        	{
-        		// コンボを元に戻す
-        		sd.MaxCombo = msd.MaxCombo;
-        	}
+        if (!mGateSetting.OverWriteLowerScores) {
+            // スコアが低かったら
+            if (scoreData.Score < localScoreData.Score) {
+                // スコアを元に戻す
+                scoreData.Score = localScoreData.Score;
+                scoreData.Rank = localScoreData.Rank;
+            }
+            // コンボが低かったら
+            if (scoreData.MaxCombo < localScoreData.MaxCombo) {
+                // コンボを元に戻す
+                scoreData.MaxCombo = localScoreData.MaxCombo;
+            }
         	/*
         	// 元の値がAAA
         	if(msd.Rank == MusicRank.AAA)
@@ -692,225 +608,194 @@ public class DialogFromGateSou {
         		}
         	}
         	*/
-        	// 元の値がMFC
-        	if(msd.FullComboType == FullComboType.MerverousFullCombo)
-        	{
-        		// MFCにする
-        		sd.FullComboType = msd.FullComboType;
-        	}
-        	// 元の値がPFC
-        	else if(msd.FullComboType == FullComboType.PerfectFullCombo)
-        	{
-        		// 取得した値がMFCでない
-        		if(sd.FullComboType != FullComboType.MerverousFullCombo)
-        		{
-        			// PFCにする
-        			sd.FullComboType = msd.FullComboType;
-        		}
-        	}
-        	// 元の値がFC
-        	else if(msd.FullComboType == FullComboType.FullCombo)
-        	{
-        		// 取得した値がMFCでもPFCでもない
-        		if(sd.FullComboType != FullComboType.MerverousFullCombo && sd.FullComboType != FullComboType.PerfectFullCombo)
-        		{
-        			// FCにする
-        			sd.FullComboType = msd.FullComboType;
-        		}
-        	}
-        	// 元の値がGFC
-        	else if(msd.FullComboType == FullComboType.GoodFullCombo)
-        	{
-        		// 取得した値がMFCでもPFCでもFCでもない
-        		if(sd.FullComboType != FullComboType.MerverousFullCombo && sd.FullComboType != FullComboType.PerfectFullCombo && sd.FullComboType != FullComboType.FullCombo)
-        		{
-        			// GFCにする
-        			sd.FullComboType = msd.FullComboType;
-        		}
-        	}
-        	// 元の値がその他
-        	else
-        	{
-        		// 取得した値がMFCでもPFCでもFCでもGFCでもない
-        		if(sd.FullComboType != FullComboType.MerverousFullCombo && sd.FullComboType != FullComboType.PerfectFullCombo && sd.FullComboType != FullComboType.FullCombo && sd.FullComboType != FullComboType.GoodFullCombo)
-        		{
-        			// 元の値にもどす
-        			sd.FullComboType = msd.FullComboType;
-        		}
-        	}
+            // 元の値がMFC
+            if (localScoreData.FullComboType == FullComboType.MerverousFullCombo) {
+                // MFCにする
+                scoreData.FullComboType = localScoreData.FullComboType;
+            }
+            // 元の値がPFC
+            else if (localScoreData.FullComboType == FullComboType.PerfectFullCombo) {
+                // 取得した値がMFCでない
+                if (scoreData.FullComboType != FullComboType.MerverousFullCombo) {
+                    // PFCにする
+                    scoreData.FullComboType = localScoreData.FullComboType;
+                }
+            }
+            // 元の値がFC
+            else if (localScoreData.FullComboType == FullComboType.FullCombo) {
+                // 取得した値がMFCでもPFCでもない
+                if (scoreData.FullComboType != FullComboType.MerverousFullCombo && scoreData.FullComboType != FullComboType.PerfectFullCombo) {
+                    // FCにする
+                    scoreData.FullComboType = localScoreData.FullComboType;
+                }
+            }
+            // 元の値がGFC
+            else if (localScoreData.FullComboType == FullComboType.GoodFullCombo) {
+                // 取得した値がMFCでもPFCでもFCでもない
+                if (scoreData.FullComboType != FullComboType.MerverousFullCombo && scoreData.FullComboType != FullComboType.PerfectFullCombo && scoreData.FullComboType != FullComboType.FullCombo) {
+                    // GFCにする
+                    scoreData.FullComboType = localScoreData.FullComboType;
+                }
+            }
+            // 元の値がその他
+            else {
+                // 取得した値がMFCでもPFCでもFCでもGFCでもない
+                if (scoreData.FullComboType != FullComboType.MerverousFullCombo && scoreData.FullComboType != FullComboType.PerfectFullCombo && scoreData.FullComboType != FullComboType.FullCombo && scoreData.FullComboType != FullComboType.GoodFullCombo) {
+                    // 元の値にもどす
+                    scoreData.FullComboType = localScoreData.FullComboType;
+                }
+            }
         }
-        switch(mPattern)
-        {
+        switch (mPattern) {
             case bSP:
-            	ms.bSP = sd;
-            	break;
+                musicScore.bSP = scoreData;
+                break;
             case BSP:
-            	ms.BSP = sd;
-            	break;
+                musicScore.BSP = scoreData;
+                break;
             case DSP:
-            	ms.DSP = sd;
-            	break;
+                musicScore.DSP = scoreData;
+                break;
             case ESP:
-            	ms.ESP = sd;
-            	break;
+                musicScore.ESP = scoreData;
+                break;
             case CSP:
-            	ms.CSP = sd;
-            	break;
+                musicScore.CSP = scoreData;
+                break;
             case BDP:
-            	ms.BDP = sd;
-            	break;
+                musicScore.BDP = scoreData;
+                break;
             case DDP:
-            	ms.DDP = sd;
-            	break;
+                musicScore.DDP = scoreData;
+                break;
             case EDP:
-            	ms.EDP = sd;
-            	break;
+                musicScore.EDP = scoreData;
+                break;
             case CDP:
-            	ms.CDP = sd;
-            	break;
+                musicScore.CDP = scoreData;
+                break;
         }
-        mScoreList.put(mItemId, ms);
+        mScoreList.put(mItemId, musicScore);
         FileReader.saveScoreData(mParent, mRivalId, mScoreList);
         DecimalFormat df = new DecimalFormat("0,000,000");
         UniquePattern c = mSouList.get(mCurrentPage);
-        String toastString = 
-        		(mRivalName==null?"":"Rival: "+mRivalName+"\n") + 
-        		mPattern.toString() + " : " + c.musics.get(mItemId).Name+"\n"+
-        		"  Full Combo :  "+(msd.FullComboType.equals(sd.FullComboType)?"":msd.FullComboType.toString()+" -> ")+sd.FullComboType.toString()+"\n"+
-        		"  Rank :  "+(msd.Rank.equals(sd.Rank)?"":(msd.Rank.toString()+" -> "))+sd.Rank.toString()+"\n"+
-        		"  Score :  "+(msd.Score==sd.Score?"":(df.format(msd.Score)+" -> "))+df.format(sd.Score)+"\n"+
-        		"Max Combo :  "+(msd.MaxCombo==sd.MaxCombo?"":(String.valueOf(msd.MaxCombo)+" -> "))+String.valueOf(sd.MaxCombo)+"\n"+
-        		"Play Count:  "+(msd.PlayCount==sd.PlayCount?String.valueOf(sd.ClearCount)+"/"+String.valueOf(sd.PlayCount):String.valueOf(msd.ClearCount)+"/"+String.valueOf(msd.PlayCount)+" -> "+String.valueOf(sd.ClearCount)+"/"+String.valueOf(sd.PlayCount));
+        String toastString =
+                (mRivalName == null ? "" : "Rival: " + mRivalName + "\n") +
+                        mPattern.toString() + " : " + c.musics.get(mItemId).Name + "\n" +
+                        "  Full Combo :  " + (localScoreData.FullComboType.equals(scoreData.FullComboType) ? "" : localScoreData.FullComboType + " -> ") + scoreData.FullComboType.toString() + "\n" +
+                        "  Rank :  " + (localScoreData.Rank.equals(scoreData.Rank) ? "" : (localScoreData.Rank + " -> ")) + scoreData.Rank.toString() + "\n" +
+                        "  Score :  " + (localScoreData.Score == scoreData.Score ? "" : (df.format(localScoreData.Score) + " -> ")) + df.format(scoreData.Score) + "\n" +
+                        "Max Combo :  " + (localScoreData.MaxCombo == scoreData.MaxCombo ? "" : (localScoreData.MaxCombo + " -> ")) + scoreData.MaxCombo + "\n" +
+                        "Play Count:  " + (localScoreData.PlayCount == scoreData.PlayCount ? scoreData.ClearCount + "/" + scoreData.PlayCount : localScoreData.ClearCount + "/" + localScoreData.PlayCount + " -> " + scoreData.ClearCount + "/" + scoreData.PlayCount);
         mLogView2.setText(toastString);
         return true;
-	}
+    }
 
-	private int mRetryCount = 0;
-	@android.webkit.JavascriptInterface
+    private int mRetryCount = 0;
+
+    @android.webkit.JavascriptInterface
     public void viewSource(final String src) {
-        mHandler.post(new Runnable() {
-            public void run() {
-            	mWebProgress.setProgress(0);
-            	
-            	if(mCanceled)
-            	{
-            		return;
-            	}
-            	
-            	try
-            	{
-                	if(!analyzeScore(src))
-                	{
-                		String toastString;
-                		switch(TextUtil.checkLoggedIn(src))
-                		{
-                			// ログイン済みエラーなし
-	                		case 0:
-	                            toastString = 
-                        		"Complete !!\n\n"+
-                        		(mRivalName==null?"":"Rival:  "+mRivalName+"\n")+
-                        		"No Data...";
-	                            mLogView2.setText(toastString);
-                            	mRetryCount = 0;
-	                			break;
-	                		// ログインしていない
-	                		case 1:
-	                	        Intent intent=new Intent();
-	                	        intent.setClassName("jp.linanfine.dsma","jp.linanfine.dsma.activity.GateLogin");
-	                	        
-	                	        cancel();
-	                	 
-	                	        mParent.startActivityForResult(intent, LoginRequestCode);
-	                	        mRetryCount = 0;
-	                            (new Thread(new Runnable() {
-	                                public void run() {
-	                            		try { Thread.sleep(1000);} catch (InterruptedException e) {}
-	                                    mHandler.post(new Runnable() {
-	                                    	public void run() {
-	                                            mDialog.cancel();
-	                                    	}
-	                                    });
-	                                    }
-	                                }
-	                            )).start();
-	                    		return;
-	                		// 不明
-	                		case -1:
-	                    		if(mRetryCount < 3)
-	                    		{
-	                                toastString = 
-	                                		"Retying...";
-	                                Toast.makeText(mParent, toastString, Toast.LENGTH_LONG).show();
-	                                ++mRetryCount;
-	                                --mCurrentPage;
-	                    		}
-	                    		else
-	                    		{
-	                    			Toast.makeText(mParent, mParent.getResources().getString(R.string.dialog_networkerrorexit), Toast.LENGTH_LONG).show();
-	                    		    mDialog.cancel();
-	                    		    return;
-	                    		}
-	                			break;
-                		}
-                	}
-                	else
-                	{
-                		mRetryCount = 0;
-                	}
-            	}
-            	catch(Exception e)
-            	{
-            		return;
-            	}
-            	
-            	try { Thread.sleep(3000);} catch (InterruptedException e) {}
-            	++mCurrentPage;
-            	mProgress.setProgress(mCurrentPage*100);
-            	mCurrent.setText(String.valueOf(mCurrentPage));
-            	mPercent.setText(String.valueOf(mPageCount==0?0:(100*mCurrentPage/mPageCount))+"%");
-            	if(mCurrentPage >= mPageCount)
-            	{
-                	mLogView.setText("Finish.");
-            		Toast.makeText(mParent, "Finish.", Toast.LENGTH_LONG).show();
-                    (new Thread(new Runnable() {
-                        public void run() {
-                    		try { Thread.sleep(1000);} catch (InterruptedException e) {}
-                            mHandler.post(new Runnable() {
-                            	public void run() {
-                                    mDialog.cancel();
-                            	}
-                            });
-                            }
-                        }
-                    )).start();
-            		return;
-            	}
-                UniquePattern c = mSouList.get(mCurrentPage);
-                mItemId = c.MusicId;
-                if(!mWebMusicIds.containsKey(mItemId) || mWebMusicIds.get(mItemId).idOnWebPage == "")
-                {
-            		new AlertDialog.Builder(mParent)
-        	        .setIcon(android.R.drawable.ic_dialog_alert)
-        	        .setMessage(mParent.getResources().getString(R.string.illegal_id_alert))
-        	        .setCancelable(true)
-        	        .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), new DialogInterface.OnClickListener() {
-        	            public void onClick(DialogInterface dialog, int whichButton) {
-        	            	
-        	            }
-        	        }).show();
-            		mDialog.cancel();
-                	return;
-                }
-                else
-                {
-                	mWebItemId = mWebMusicIds.get(mItemId).idOnWebPage;
-                }
-                mPattern = c.Pattern;
-                mRequestUri = mUriH+String.valueOf(mWebItemId)+mUriM+String.valueOf(getPatternInt(mPattern))+mUriF;
-                mLogView.setText("Loading...\n"+(mRivalName==null?"My Score\n":("Rival: "+mRivalName+"\n")) + mPattern.toString() + " : " + c.musics.get(mItemId).Name);
-                mWebView.loadUrl(mRequestUri);
+        mHandler.post(() -> {
+            mWebProgress.setProgress(0);
+
+            if (mCanceled) {
+                return;
             }
+
+            try {
+                if (!_analyzeScore(src)) {
+                    String toastString;
+                    switch (TextUtil.checkLoggedIn(src)) {
+                        // ログイン済みエラーなし
+                        case 0:
+                            toastString =
+                                    "Complete !!\n\n" +
+                                            (mRivalName == null ? "" : "Rival:  " + mRivalName + "\n") +
+                                            "No Data...";
+                            mLogView2.setText(toastString);
+                            mRetryCount = 0;
+                            break;
+                        // ログインしていない
+                        case 1:
+                            Intent intent = new Intent();
+                            intent.setClassName("jp.linanfine.dsma", "jp.linanfine.dsma.activity.GateLogin");
+
+                            cancel();
+
+                            mParent.startActivityForResult(intent, LoginRequestCode);
+                            mRetryCount = 0;
+                            (new Thread(() -> {
+                                try {
+                                    Thread.sleep(1000);
+                                } catch (InterruptedException ignored) {
+                                }
+                                mHandler.post(() -> mDialog.cancel());
+                            }
+                            )).start();
+                            return;
+                        // 不明
+                        case -1:
+                            if (mRetryCount < 3) {
+                                toastString =
+                                        "Retying...";
+                                Toast.makeText(mParent, toastString, Toast.LENGTH_LONG).show();
+                                ++mRetryCount;
+                                --mCurrentPage;
+                            } else {
+                                Toast.makeText(mParent, mParent.getResources().getString(R.string.dialog_networkerrorexit), Toast.LENGTH_LONG).show();
+                                mDialog.cancel();
+                                return;
+                            }
+                            break;
+                    }
+                } else {
+                    mRetryCount = 0;
+                }
+            } catch (Exception e) {
+                return;
+            }
+
+            try {
+                Thread.sleep(3000);
+            } catch (InterruptedException ignored) {
+            }
+            ++mCurrentPage;
+            mProgress.setProgress(mCurrentPage * 100);
+            mCurrent.setText(String.valueOf(mCurrentPage));
+            mPercent.setText((mPageCount == 0 ? 0 : (100 * mCurrentPage / mPageCount)) + "%");
+            if (mCurrentPage >= mPageCount) {
+                mLogView.setText("Finish.");
+                Toast.makeText(mParent, "Finish.", Toast.LENGTH_LONG).show();
+                (new Thread(() -> {
+                    try {
+                        Thread.sleep(1000);
+                    } catch (InterruptedException ignored) {
+                    }
+                    mHandler.post(() -> mDialog.cancel());
+                }
+                )).start();
+                return;
+            }
+            UniquePattern c = mSouList.get(mCurrentPage);
+            mItemId = c.MusicId;
+            if (!mWebMusicIds.containsKey(mItemId) || mWebMusicIds.get(mItemId).idOnWebPage.equals("")) {
+                new AlertDialog.Builder(mParent)
+                        .setIcon(android.R.drawable.ic_dialog_alert)
+                        .setMessage(mParent.getResources().getString(R.string.illegal_id_alert))
+                        .setCancelable(true)
+                        .setPositiveButton(mParent.getResources().getString(R.string.strings_global____ok), (dialog, whichButton) -> {
+
+                        }).show();
+                mDialog.cancel();
+                return;
+            } else {
+                mWebItemId = mWebMusicIds.get(mItemId).idOnWebPage;
+            }
+            mPattern = c.Pattern;
+            mRequestUri = mUriH + mWebItemId + mUriM + getPatternInt(mPattern) + mUriF;
+            mLogView.setText("Loading...\n" + (mRivalName == null ? "My Score\n" : ("Rival: " + mRivalName + "\n")) + mPattern.toString() + " : " + c.musics.get(mItemId).Name);
+            mWebView.loadUrl(mRequestUri);
         });
     }
-	
+
 }
